[SPARK-40467][SS] Split FlatMapGroupsWithState down to multiple test suites
